# docs/metis-api.openapi.yaml
openapi: 3.1.0
info:
  title: METIS Execution Plane API
  version: 2.0.0
  description: |
    AI-native decision platform API.
    Execution plane: deterministic strategy execution, artifact registry, approval workflow.
  contact:
    name: METIS Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://metis.example.com/api
    description: Production
components:
  schemas:
...
    ReplayResult:
      type: object
      description: The outcome of re-executing a historical decision.
      required:
        [identical, decisionId, replayedAt, artifactVersion, originalWinner,
         replayedWinner, originalChainHash, replayedChainHash, diff]
      properties:
        identical:
          type: boolean
          description: True when the two chain hashes match exactly
        decisionId:
          type: string
        replayedAt:
          type: string
          format: date-time
        artifactVersion:
          type: string
        originalWinner:
          type: [string, 'null']
        replayedWinner:
          type: [string, 'null']
        originalChainHash:
          type: string
        replayedChainHash:
          type: string
        diff:
          type: array
          description: Empty when identical is true
          items:
            type: object
            required: [path]
            properties:
              path:
                type: string
              original: {}
              replayed: {}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /decisions:
    post:
      summary: Make a decision
      description: |
        Execute a strategy against a request and return the trace.

        Integrations are resolved before the deterministic core runs and their
        values arrive in `request.input`; this operation does not fetch them.
        A caller that wants resolution asks for it by supplying the fields, or
        by calling a service configured with a gateway.

        `occurredAt` is required and is never defaulted to the current time.
        Defaulting it would make a decision depend on when it was made, which
        is the one thing replay cannot tolerate.

        Two implementations serve this: the console's development store, and
        the JVM service in `engines/kotlin`. Both produce the same chain hash
        for the same input — see docs/conformance/service-cases.json.
      operationId: executeDecision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [artifactId, request]
              properties:
                artifactId:
                  type: string
                request:
                  type: object
                  required: [tenantId, customerId, channel, placement, occurredAt, input]
                  properties:
                    tenantId:
                      type: string
                    customerId:
                      type: string
                    channel:
                      type: string
                    placement:
                      type: string
                    occurredAt:
                      type: string
                      format: date-time
                      description: An input, never the clock.
                    input:
                      type: object
                      description: Customer and context attributes, already resolved.
                    contactHistory:
                      type: object
                    consent:
                      type: object
      responses:
        '200':
          description: The decision and its trace
          content:
            application/json:
              schema:
                type: object
                required: [id, decision, chainHash]
                properties:
                  id:
                    type: string
                    description: Content-addressed. The first 16 characters of the chain hash.
                  decision:
                    type: object
                    description: The reproducible half, exactly as hashed.
                  chainHash:
                    type: string
        '400':
          description: Malformed request, or an artifact the engine cannot execute
        '404':
          description: No such artifact
...
  /decisions/{decisionId}/replay:
    post:
      summary: Re-execute a historical decision and compare it to the original
      operationId: replayDecision
      parameters:
        - name: decisionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |
            The replay outcome. `identical` is the claim the platform is making;
            when it is false, `diff` says exactly where reproduction broke.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplayResult'
        '404':
          description: No such decision

  # ---------------------------------------------------------------------------
  # Governance
  # ---------------------------------------------------------------------------
...